HBX-2182: Review and adapt the tests to remove the stack traces caused by the dialect and/or connection not being specified - Rename class 'org.hibernate.tools.test.util.DummyConnectionProvider' to 'org.hibernate.tools.test.util.internal.ConnectionProvider'
